+++ dwds/lib/src/loaders/legacy.dart
@@ -80,6 +80,9 @@ class LegacyStrategy extends LoadStrategy {
   String get moduleFormat => 'ddc';
 
   @override
+  String get loadLibrariesModule => 'dart_library.ddk.js';
+
+  @override
   String get loadLibrariesSnippet =>
       'for(let module of dart_library.libraries()) {\n'
       'dart_library.import(module)[module];\n'