+++ fixtures/_testPackage/web/main.dart
@@ -31,7 +31,7 @@ void main() {
   // for evaluation
   Timer.periodic(const Duration(seconds: 1), (_) {
     printLocal();
-    printFieldFromLibraryClass();
+    printFieldFromLibraryClass(); // Breakpoint: callPrintFieldFromLibraryClass
     printFieldFromLibraryPartClass();
     printFieldMain();
     printGlobal();
@@ -53,7 +53,7 @@ void printLocal() {
 }
 
 void printFieldFromLibraryClass() {
-  var instance = TestLibraryClass(1, 2);
+  var instance = TestLibraryClass(1, 2); // Breakpoint: createLibraryObject
   print('$instance'); // Breakpoint: printFieldFromLibraryClass
 }
 
@@ -109,7 +109,7 @@ void printNestedObjectsMultiLine() {
 }
 
 void printObjectMultiLine() {
-  print(// Breakpoint: printMultiLine
+  print( // Breakpoint: printMultiLine
     createObject() // Breakpoint: printObjectMultiLine
       ..initialize(),
   );